Refactored GraphStore by extracting addSorted functions
