+++ argo-scholar/src/stores/GraphStore.js
@@ -17,9 +17,9 @@ import {
   Intent,
 } from "@blueprintjs/core";
 import { Frame, graph } from "../graph-frontend";
-import { toaster } from "../notifications/client";
 import pageRank from "ngraph.pagerank";
 import appState from ".";
+import * as BackendAPIUtils from "../services/BackendAPIUtils";
 
 export default class GraphStore {
   initialGlobalConfig = {
@@ -534,487 +534,6 @@ export default class GraphStore {
     }
   }
 
-  /**
-   * Adding sorted citations from Semantic Scholar's backend API through multiple steps.
-   * Step 1: fetch multiple times in the for-loop and store the node id arrays in the promises array since maximum pageSize is 10 in order for the API to return ordered results.
-   * Step 2: use Promise.all() to concatenate all node ids in one array.
-   * Step 3: pick the first 5 non-duplicate ids and fetch node info with Semantic Scholar's public API.
-   * Backup Step: if backend API fails, call addRandomCitations() to add 5 unsorted nodes with Semantic Scholar's public API.
-   * @param {*} sortMethod
-   */
-  addSortedCitations(sortMethod) {
-    // Fetch request body and params
-    let curcount = 0;
-    let pageSize = 10;
-    let offset = 0;
-    // Change this to query more pages
-    let pageLimit = 3;
-
-    const rightClickedNodeId =
-      this.frame.rightClickedNode.data.ref.id.toString();
-    var requestURL =
-      "https://www.semanticscholar.org/api/1/search/paper/" +
-      rightClickedNodeId +
-      "/citations";
-
-    let promises = [];
-    // 30 ids in total, 10 each iteration
-    for (let page = 1; page <= pageLimit; page++) {
-      var requestBody = JSON.stringify({
-        page: page,
-        pageSize: pageSize,
-        sort: sortMethod,
-        authors: [],
-        coAuthors: [],
-        venues: [],
-        yearFilter: null,
-        requireViewablePdf: false,
-        publicationTypes: [],
-        externalContentTypes: [],
-        fieldsOfStudy: [],
-      });
-      promises.push(
-        fetch(requestURL, {
-          method: "post",
-          headers: {
-            Accept: "application/json, text/plain, */*",
-            "Content-Type": "application/json",
-          },
-          body: requestBody,
-        })
-          .then((res) => {
-            if (res.ok) {
-              return res.json();
-            } else {
-              throw "Connection to backend API failed.";
-            }
-          })
-          // Mapping ids onto an array
-          .then((citations) => {
-            var PaperIds = citations.results.map(function (citation) {
-              return citation.id;
-            });
-            return PaperIds;
-          })
-      );
-    }
-
-    Promise.all(promises)
-      .then((result) => {
-        // Concatenate the ids into one long array
-        let PaperIds = result.flat();
-        console.log(PaperIds);
-        console.log("Node array length: " + PaperIds.length);
-        // Creating nodes
-        let edgesArr = toJS(appState.graph.rawGraph.edges);
-        let addedNodes = [];
-        let fetches = [];
-
-        let duplicateNodes = 0;
-        for (let i = 0; i < PaperIds.length; i++) {
-          // Null check
-          if (!PaperIds[i]) {
-            continue;
-          }
-
-          // Checking duplicate nodes
-          if (
-            edgesArr.some(
-              (edge) =>
-                edge.source_id == PaperIds[i] &&
-                edge.target_id == rightClickedNodeId
-            )
-          ) {
-            console.log("Already containing nodes!");
-            duplicateNodes += 1;
-            continue;
-          }
-
-         
-          curcount += 1;
-
-          // Fetching node info from public API
-          let citationAPI =
-            "https://api.semanticscholar.org/v1/paper/" + PaperIds[i];
-          fetches.push(
-            fetch(citationAPI)
-              .then((res) => {
-                if (res.ok) {
-                  return res.json();
-                } else {
-                  throw "Connection to public API failed.";
-                }
-              })
-              .then((response) => {
-                this.addNodetoGraph(response, rightClickedNodeId, 0);
-                appState.graph.selectedNodes = [];
-                appState.graph.frame.selection = [];
-                this.frame.addFrontEndNodeInARow(
-                  rightClickedNodeId,
-                  response.paperId,
-                  offset,
-                  0
-                );
-                let paperNode = appState.graph.process.graph.getNode(
-                  response.paperId
-                );
-                addedNodes.push(paperNode);
-                paperNode.renderData.textHolder.children[0].element.override = true;
-                offset += 1;
-              })
-              .catch(function (err) {
-                console.warn(
-                  "Cannot fetch node info through Semantic Scholar's public API. " +
-                    requestURL,
-                  err
-                );
-              })
-          );
-
-          // Exit condition
-          if (curcount >= 5) {
-            break;
-          }
-        }
-        console.log("Graph contains " + duplicateNodes + " duplicate nodes.");
-
-        // Check for remaining nodes
-        if(PaperIds.length<pageLimit*pageSize &&PaperIds.length==duplicateNodes ){
-          console.log("All citations added");
-          toaster.show({
-            message:
-              "All the citations for this node have been added.",
-            intent: Intent.WARNING,
-          });
-        }
-
-        // Check for upper limit
-        if (duplicateNodes >=pageLimit*pageSize) {
-          console.log("Limit reached");
-          toaster.show({
-            message:
-              "Cannot exceed maximum number of citations allowed per node (30).",
-            intent: Intent.WARNING,
-          });
-        }
-
-        Promise.all(fetches).then(() => {
-          addedNodes.forEach((node) => {
-            appState.graph.selectedNodes.push(node);
-            appState.graph.frame.selection.push(node);
-          });
-        });
-      })
-      // Error message, using backup method instead
-      .catch((error) => {
-        console.error(
-          "Error occurred when requesting sorted citations through Semantic Scholar's backend API. Adding random citations. Reason:",
-          error
-        );
-        toaster.show({
-          message:
-            "Error getting sorted results. Adding unsorted results from Semantic Scholar instead.",
-          intent: Intent.WARNING,
-        });
-        this.addRandomCitations();
-      });
-
-    appState.graph.frame.updateNodesShowingLabels();
-  }
-
-  //Backup: add random citaions to right clicked node
-  addRandomCitations() {
-    const rightClickedNodeId =
-      this.frame.rightClickedNode.data.ref.id.toString();
-    let curcount = 0;
-    let offset = 0;
-    let apiurl =
-      "https://api.semanticscholar.org/v1/paper/" + rightClickedNodeId;
-    fetch(apiurl)
-      .then((res) => {
-        return res.json();
-      })
-      .then((response) => {
-        return response.citations;
-      })
-      .then((citations) => {
-        console.log(citations);
-        let edgesArr = toJS(appState.graph.rawGraph.edges);
-        let addedNodes = [];
-        let fetches = [];
-        for (let i = 0; i < citations.length; i++) {
-          if (
-            edgesArr.some(
-              (edge) =>
-                edge.source_id == citations[i].paperId &&
-                edge.target_id == rightClickedNodeId
-            )
-          ) {
-            console.log("contains key");
-            continue;
-          }
-          curcount += 1;
-          let citationAPI =
-            "https://api.semanticscholar.org/v1/paper/" + citations[i].paperId;
-          fetches.push(
-            fetch(citationAPI)
-              .then((res) => {
-                if (res.ok) {
-                  return res.json();
-                } else {
-                  throw "Connection failed.";
-                }
-              })
-              .then((response) => {
-                this.addNodetoGraph(response, rightClickedNodeId, 0);
-                appState.graph.selectedNodes = [];
-                appState.graph.frame.selection = [];
-                this.frame.addFrontEndNodeInARow(
-                  rightClickedNodeId,
-                  response.paperId,
-                  offset,
-                  0
-                );
-                let paperNode = appState.graph.process.graph.getNode(
-                  response.paperId
-                );
-                addedNodes.push(paperNode);
-                paperNode.renderData.textHolder.children[0].element.override = true;
-                offset += 1;
-                // appState.graph.process.graph.getNode(
-                //   response.paperId
-                // ).renderData.textHolder.children[0].element.override = true;
-                // offset += 1;
-                // appState.graph.selectedNodes = [];
-                // appState.graph.frame.selection = [];
-              })
-          );
-          if (curcount >= 5) {
-            break;
-          }
-        }
-        console.log("Successfully added random citation nodes.");
-        Promise.all(fetches).then(() => {
-          addedNodes.forEach((node) => {
-            appState.graph.selectedNodes.push(node);
-            appState.graph.frame.selection.push(node);
-          });
-        });
-      })
-      .catch((error) => {
-        console.error("Cannot add random nodes. Reason:", error);
-        toaster.show({
-          message: "Cannot load results. Please check network connection.",
-          intent: Intent.DANGER,
-          iconName: "warning-sign",
-        });
-      });
-    appState.graph.frame.updateNodesShowingLabels();
-  }
-
-  //add sorted references to right clicked node
-  addSortedReferences(sortMethod) {
-    let curcount = 0;
-    let pageSize = 10;
-    let offset = 0;
-
-    const rightClickedNodeId =
-      this.frame.rightClickedNode.data.ref.id.toString();
-    var requestURL =
-      "https://argo-cors-anywhere.herokuapp.com/https://www.semanticscholar.org/api/1/paper/" +
-      rightClickedNodeId +
-      "/citations?sort=" +
-      sortMethod +
-      "&offset=" +
-      offset +
-      "&citationType=citedPapers&citationsPageSize=" +
-      pageSize;
-
-    fetch(requestURL, {
-      method: "get",
-      headers: {
-        "Content-Type": "application/json",
-      },
-    })
-      .then((res) => {
-        if (res.ok) {
-          return res.json();
-        } else {
-          throw "Connection failed.";
-        }
-      })
-      .then((references) => {
-        var PaperIds = references.citations.map(function (reference) {
-          return reference.id;
-        });
-
-        let edgesArr = toJS(appState.graph.rawGraph.edges);
-        let addedNodes = [];
-        let fetches = [];
-        console.log(PaperIds);
-        for (let i = 0; i < PaperIds.length; i++) {
-          if (!PaperIds[i]) {
-            continue;
-          }
-          if (
-            edgesArr.some(
-              (edge) =>
-                edge.source_id == PaperIds[i] &&
-                edge.target_id == rightClickedNodeId
-            )
-          ) {
-            console.log("Already containing keys!");
-            continue;
-          }
-          curcount += 1;
-          let citationAPI =
-            "https://api.semanticscholar.org/v1/paper/" + PaperIds[i];
-          fetches.push(
-            fetch(citationAPI)
-              .then((res) => {
-                if (res.ok) {
-                  return res.json();
-                } else {
-                  throw "Connection failed.";
-                }
-              })
-              .then((response) => {
-                this.addNodetoGraph(response, rightClickedNodeId, 1);
-                appState.graph.selectedNodes = [];
-                appState.graph.frame.selection = [];
-                this.frame.addFrontEndNodeInARow(
-                  rightClickedNodeId,
-                  response.paperId,
-                  offset,
-                  1
-                );
-                let paperNode = appState.graph.process.graph.getNode(
-                  response.paperId
-                );
-                addedNodes.push(paperNode);
-                paperNode.renderData.textHolder.children[0].element.override = true;
-                offset += 1;
-              })
-              .catch(function (err) {
-                console.warn(
-                  "Error adding paper nodes through Semantic Scholar API. " +
-                    requestURL,
-                  err
-                );
-              })
-          );
-          if (curcount >= 5) {
-            break;
-          }
-        }
-        console.log("Sorted references added!");
-        Promise.all(fetches).then((response) => {
-          addedNodes.forEach((node) => {
-            appState.graph.selectedNodes.push(node);
-            appState.graph.frame.selection.push(node);
-          });
-        });
-      })
-
-      .catch((error) => {
-        console.error(
-          "Error occurred when requesting sorted references through web API. Adding random references. Reason:",
-          error
-        );
-        toaster.show({
-          message:
-            "Error getting sorted results. Adding unsorted results from Semantic Scholar instead.",
-          intent: Intent.WARNING,
-        });
-        this.addRandomReferences();
-      });
-    appState.graph.frame.updateNodesShowingLabels();
-  }
-
-  //Backup: add random references to the right clicked node
-  addRandomReferences() {
-    const rightClickedNodeId =
-      this.frame.rightClickedNode.data.ref.id.toString();
-    let curcount = 0;
-    let offset = 0;
-    let apiurl =
-      "https://api.semanticscholar.org/v1/paper/" + rightClickedNodeId;
-    fetch(apiurl)
-      .then((res) => {
-        return res.json();
-      })
-      .then((response) => {
-        return response.references;
-      })
-      .then((references) => {
-        console.log(references);
-        let edgesArr = toJS(appState.graph.rawGraph.edges);
-        let addedNodes = [];
-        let fetches = [];
-        for (let i = 0; i < references.length; i++) {
-          if (
-            edgesArr.some(
-              (edge) =>
-                edge.source_id == rightClickedNodeId &&
-                edge.target_id == references[i].paperId
-            )
-          ) {
-            console.log("contains key");
-            continue;
-          }
-          curcount += 1;
-          let citationAPI =
-            "https://api.semanticscholar.org/v1/paper/" + references[i].paperId;
-          fetches.push(
-            fetch(citationAPI)
-              .then((res) => {
-                if (res.ok) {
-                  return res.json();
-                } else {
-                  throw "Connection failed.";
-                }
-              })
-              .then((response) => {
-                this.addNodetoGraph(response, rightClickedNodeId, 1);
-                appState.graph.selectedNodes = [];
-                appState.graph.frame.selection = [];
-                this.frame.addFrontEndNodeInARow(
-                  rightClickedNodeId,
-                  response.paperId,
-                  offset,
-                  1
-                );
-                let paperNode = appState.graph.process.graph.getNode(
-                  response.paperId
-                );
-                addedNodes.push(paperNode);
-                paperNode.renderData.textHolder.children[0].element.override = true;
-                offset += 1;
-              })
-          );
-          if (curcount >= 5) {
-            break;
-          }
-        }
-        console.log("Successfully added random reference nodes.");
-        Promise.all(fetches).then((response) => {
-          addedNodes.forEach((node) => {
-            appState.graph.selectedNodes.push(node);
-            appState.graph.frame.selection.push(node);
-          });
-        });
-      })
-      .catch((error) => {
-        console.error("Cannot add random nodes. Reason:", error);
-        toaster.show({
-          message: "Cannot load results. Please check network connection.",
-          intent: Intent.DANGER,
-          iconName: "warning-sign",
-        });
-      });
-
-    appState.graph.frame.updateNodesShowingLabels();
-  }
   setUpFrame() {
     if (this.frame) {
       this.frame.getGraph().clear();
@@ -1089,28 +608,28 @@ export default class GraphStore {
               children: [
                 MenuItemFactory({
                   onClick: () => {
-                    this.addSortedCitations("relevance");
+                    BackendAPIUtils.addSortedCitations("relevance", this);
                   },
                   text: "Sort By Relevance",
                   key: "Sort By Relevance",
                 }),
                 MenuItemFactory({
                   onClick: () => {
-                    this.addSortedCitations("is-influential");
+                    BackendAPIUtils.addSortedCitations("is-influential", this);
                   },
                   text: "Sort By Most Infleunced Papers",
                   key: "Sort By Most Infleunced Papers",
                 }),
                 MenuItemFactory({
                   onClick: () => {
-                    this.addSortedCitations("total-citations");
+                    BackendAPIUtils.addSortedCitations("total-citations", this);
                   },
                   text: "Sort By Citation Count",
                   key: "Sort By Citation Count",
                 }),
                 MenuItemFactory({
                   onClick: () => {
-                    this.addSortedCitations("pub-date");
+                    BackendAPIUtils.addSortedCitations("pub-date", this);
                   },
                   text: "Sort By Recency",
                   key: "Sort By Recency",
@@ -1124,21 +643,21 @@ export default class GraphStore {
               children: [
                 MenuItemFactory({
                   onClick: () => {
-                    this.addSortedReferences("relevance");
+                    BackendAPIUtils.addSortedReferences("relevance", this);
                   },
                   text: "Sort By Relevance",
                   key: "Sort By Relevance",
                 }),
                 MenuItemFactory({
                   onClick: () => {
-                    this.addSortedReferences("is-influential");
+                    BackendAPIUtils.addSortedReferences("is-influential", this);
                   },
                   text: "Sort By Most Infleunced Papers",
                   key: "Sort By Most Infleunced Papers",
                 }),
                 MenuItemFactory({
                   onClick: () => {
-                    this.addSortedReferences("year");
+                    BackendAPIUtilsZ.addSortedReferences("year", this);
                   },
                   text: "Sort By Recency",
                   key: "Sort By Recency",
